requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Odoo Integration
odoorpc==0.10.0
requests==2.31.0
httpx==0.25.2

# Device Management
pyserial==3.5
pyusb==1.2.1
pybluez==0.23
netifaces==0.11.0

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
sqlite3

# Web Interface
flask==3.0.0
flask-cors==4.0.0
jinja2==3.1.2
werkzeug==3.0.1

# Configuration & Environment
pyyaml==6.0.1
python-dotenv==1.0.0
configparser==6.0.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
psutil==5.9.6

# Security
cryptography==41.0.8
pyjwt==2.8.0
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
factory-boy==3.3.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
mkdocs==1.5.3
mkdocs-material==9.4.8

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
schedule==1.2.0
